refactor(media): extract MediaUrl leaf, break MediaHelper/Generator inverted dependency ss-143
