+++ laravel/app/Helpers/MediaHelper.php
@@ -2,12 +2,12 @@
 
 namespace App\Helpers;
 
-use Illuminate\Support\Facades\Storage;
+use App\Contracts\Media\MediaUrlGeneratorInterface;
 
 class MediaHelper
 {
     /**
-     * Get absolute URL for a path using a disk from configuration.
+     * Get absolute URL for a path using a disk resolved from configuration.
      *
      * @param  string|null  $path  Relative storage path
      * @param  string  $diskConfigKey  Config key for the disk name
@@ -15,32 +15,9 @@ class MediaHelper
      */
     public static function getUrl(?string $path, string $diskConfigKey = 'ai.tts.storage.disk', string $defaultDisk = 'public'): ?string
     {
-        if (! $path) {
-            return null;
-        }
-
         $diskName = ConfigHelper::getString($diskConfigKey, $defaultDisk);
 
-        return self::toAbsolute(Storage::disk($diskName)->url($path));
-    }
-
-    /**
-     * Promote a disk-generated URL to an absolute one.
-     *
-     * Cloud disks (s3/R2) already return absolute URLs; local disks return a
-     * site-relative path, which we prefix with the app URL. Single source of
-     * truth reused by the media services and the model accessors (Game, Level).
-     *
-     * Assumes a disk URL is either an absolute http(s) URL or a scheme-less
-     * site-relative path. Protocol-relative ("//cdn/…") or other schemes are
-     * not produced by the current disks; revisit this check if a CDN that
-     * emits them is added.
-     */
-    public static function toAbsolute(string $url): string
-    {
-        return str_starts_with($url, 'http://') || str_starts_with($url, 'https://')
-            ? $url
-            : url($url);
+        return app(MediaUrlGeneratorInterface::class)->getUrl($path, $diskName);
     }
 
     /**
